backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Redis (for caching and sessions)
redis==5.0.1
hiredis==2.2.3

# Email
emails==0.6.0

# File handling
python-magic==0.4.27
Pillow==10.1.0

# Date and time
python-dateutil==2.8.2

# Validation and parsing
validators==0.22.0
phonenumbers==8.13.25

# Async support
asyncio==3.4.3
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Background tasks
celery==5.3.4
flower==2.0.1

# Data processing
pandas==2.1.3
numpy==1.25.2

# QR Code generation
qrcode[pil]==7.4.2

# Cryptography (additional)
cryptography==41.0.7

# LINE Bot SDK
line-bot-sdk==3.5.0

# OCR and AI services
opencv-python==4.8.1.78

# Banking and payment integration
stripe==7.8.0

# Excel file handling
openpyxl==3.1.2
xlsxwriter==3.1.9

# PDF generation
reportlab==4.0.7
weasyprint==60.2

# Image processing
Pillow==10.1.0

# Timezone handling
pytz==2023.3

# Environment detection
python-decouple==3.8

# Rate limiting
slowapi==0.1.9

# CORS and middleware
starlette==0.27.0

# Backup and storage
boto3==1.34.0  # For AWS S3 integration

# Monitoring
prometheus-client==0.19.0

# Health checks
psutil==5.9.6

